+++ ReleaseNotes.md
@@ -1,5 +1,22 @@
 # Worksmith release notes
 
+## 0.1.5
+Terminate a workflow or sequence from code (a programmable alternative to conditional="predicate" attribute)
+
+in execute scope `this` holds workflow execution api. you do pass your done, and not invoke it yourself.
+
+Note: in later versions you will not have to provide done
+```javascript
+//terminate a workflow
+function execute(done) {
+    this.workflow.terminate(/*err*/, /*result*/, done)
+}
+//terminate the current parent scope
+function execute(done) {
+    this.workflow.terminateParent(/*err*/, /*result*/, done)
+}
+```
+
 ## 0.1.4
 Error logging temporalily removed